M2-6128: [Mobile] Unit tests for ConstructCompletionsService.constructForFinish
