+++ src/widgets/survey/model/services/ConstructCompletionsService.ts
@@ -251,13 +251,13 @@ export class ConstructCompletionsService {
       userActions: mapUserActionsToDto(actions),
       itemIds: modifiedItemIds,
       appletEncryption: appletEncryption!,
-      flowId: flowId ?? null,
+      flowId: flowId,
       activityId: activityId,
       executionGroupKey: getExecutionGroupKey(progressRecord),
       userIdentifier: getUserIdentifier(items, recordAnswers),
       startTime: getActivityStartAt(progressRecord)!,
       endTime: getNow().getTime(),
-      scheduledTime: scheduledDate ?? undefined,
+      scheduledTime: scheduledDate,
       logActivityName: activityName,
       logCompletedAt: getNow().toUTCString(),
       client: getClientInformation(),
@@ -328,7 +328,7 @@ export class ConstructCompletionsService {
 
     this.pushToQueueService.push({
       appletId,
-      createdAt: Date.now(),
+      createdAt: getNow().getTime(),
       version: activityStorageRecord.appletVersion,
       answers: modifiedAnswers,
       userActions,
@@ -339,10 +339,10 @@ export class ConstructCompletionsService {
       executionGroupKey: getExecutionGroupKey(progressRecord),
       userIdentifier,
       startTime: getActivityStartAt(progressRecord)!,
-      endTime: Date.now(),
-      scheduledTime: scheduledDate ?? undefined,
+      endTime: getNow().getTime(),
+      scheduledTime: scheduledDate,
       logActivityName: activityName,
-      logCompletedAt: new Date().toString(),
+      logCompletedAt: getNow().toUTCString(),
       client: getClientInformation(),
       alerts,
       eventId,
