+++ src/widgets/survey/model/services/tests/ConstructCompletionsService.test.ts
@@ -1,0 +1,621 @@
+import { StoreProgress } from '@app/abstract/lib';
+import {
+  Answer,
+  Answers,
+  PipelineItem,
+  UserAction,
+} from '@app/features/pass-survey';
+import { getEmptySliderItem as getSliderItem } from '@app/features/pass-survey/model/tests/testHelpers';
+import { AnswerDto, SliderAnswerDto } from '@app/shared/api';
+import { FlowState } from '@app/widgets/survey/lib';
+import * as dateTimeUtils from '@shared/lib/utils/dateTime';
+
+import {
+  getActivityProgressMock,
+  getActivityRecordMockResult,
+  getFlowProgressMock,
+  getUserActionsMock,
+} from './testHelpers';
+import * as metaHelpers from '../../../lib/metaHelpers';
+import * as storageHelpers from '../../../lib/storageHelpers';
+import * as mappers from '../../mappers';
+import * as operations from '../../operations';
+import { ConstructCompletionsService } from '../ConstructCompletionsService';
+
+jest.mock('@app/shared/lib/services/Logger', () => ({
+  log: jest.fn(),
+  info: jest.fn(),
+  warn: jest.fn(),
+}));
+
+jest.mock('@app/features/pass-survey/model/ScoresExtractor', () => ({
+  extract: jest
+    .fn()
+    .mockReturnValue([
+      { name: 'mock-score-name-1', value: 125, flagged: true },
+    ]),
+}));
+
+jest.mock('@app/features/pass-survey/model/AlertsExtractor', () => ({
+  extractForSummary: jest.fn().mockReturnValue([
+    {
+      activityItemId: 'mock-activity-item-1',
+      message: 'mock-message-1',
+    },
+  ]),
+}));
+
+const mockNowDate = new Date(2023, 3, 8, 15, 27);
+
+jest.spyOn(dateTimeUtils, 'getNow').mockReturnValue(mockNowDate);
+
+jest.spyOn(dateTimeUtils, 'getTimezoneOffset').mockReturnValue(3);
+
+const createSvgFilesMock = jest
+  .spyOn(operations, 'createSvgFiles')
+  .mockResolvedValue([]);
+
+const saveSummaryMock = jest.fn();
+
+const mapAnswersToDtoMock = jest
+  .spyOn(mappers, 'mapAnswersToDto')
+  .mockImplementation((_: any, answers: Answers) => {
+    return [
+      convertToSliderAnswerMock(answers[0]),
+      convertToSliderAnswerMock(answers[1]),
+    ];
+  });
+
+const mapUserActionsToDtoMock = jest
+  .spyOn(mappers, 'mapUserActionsToDto')
+  .mockReturnValue([]);
+
+const getItemIdsMock = jest
+  .spyOn(operations, 'getItemIds')
+  .mockReturnValue(['mock-slider-id-1', 'mock-slider-id-2']);
+
+const mockFillNullsForHiddenItems = jest
+  .spyOn(operations, 'fillNullsForHiddenItems')
+  .mockImplementation((itemIds: string[], answers: AnswerDto[], _: any) => {
+    return { answers, itemIds };
+  });
+
+const getFlowRecordMock = jest
+  .spyOn(storageHelpers, 'getFlowRecord')
+  .mockReturnValue({
+    scheduledDate: 1245800000,
+    flowName: null,
+  } as FlowState);
+
+const getUserIdentifierMock = jest
+  .spyOn(operations, 'getUserIdentifier')
+  .mockReturnValue('mock-user-id-1');
+
+const mapAnswersToAlertsMock = jest
+  .spyOn(mappers, 'mapAnswersToAlerts')
+  .mockReturnValue([]);
+
+const getClientInformationMock = jest
+  .spyOn(metaHelpers, 'getClientInformation')
+  .mockReturnValue({
+    appId: 'mock-app-name-1',
+    appVersion: 'mock-app-version-1',
+    width: 1347,
+    height: 2358,
+  });
+
+const clearActivityStorageRecordMock = jest
+  .spyOn(storageHelpers, 'clearActivityStorageRecord')
+  .mockReturnValue();
+
+function convertToSliderAnswerMock(answer: Answer): AnswerDto {
+  return {
+    value: answer.answer as SliderAnswerDto,
+  };
+}
+
+describe('Test ConstructCompletionsService.constructForIntermediate', () => {
+  afterEach(() => {
+    jest.clearAllMocks();
+  });
+
+  it('Should push item to queue when hasSummary is false', async () => {
+    const answersMock: Answers = {
+      0: { answer: 3 },
+      1: { answer: 8 },
+    };
+
+    const itemsMock: PipelineItem[] = [
+      getSliderItem('mock-slider-item-1', 'mock-slider-id-1'),
+      getSliderItem('mock-slider-item-2', 'mock-slider-id-2'),
+    ];
+
+    const userActionsMock: UserAction[] = getUserActionsMock(answersMock);
+
+    const getActivityRecordMock = jest
+      .spyOn(storageHelpers, 'getActivityRecord')
+      .mockReturnValue(
+        getActivityRecordMockResult(answersMock, userActionsMock, itemsMock),
+      );
+
+    const pushMock = jest.fn();
+
+    const pushToQueueMock = { push: pushMock };
+
+    const dispatchMock = jest.fn();
+
+    const progress: StoreProgress = getFlowProgressMock();
+
+    const service = new ConstructCompletionsService(
+      saveSummaryMock,
+      {} as any,
+      progress,
+      pushToQueueMock,
+      dispatchMock,
+    );
+
+    //@ts-expect-error
+    service.queryDataUtils.getAppletDto = jest.fn().mockReturnValue({
+      encryption: 'applet-encryption-mock-1',
+      displayName: 'mock-applet-name-1',
+    });
+
+    await service.construct({
+      completionType: 'intermediate',
+      activityId: 'mock-activity-id-1',
+      activityName: 'mock-activity-name-1',
+      appletId: 'mock-applet-id-1',
+      eventId: 'mock-event-id-1',
+      flowId: 'mock-flow-id-1',
+      order: 0,
+    });
+
+    expect(getActivityRecordMock).toBeCalledTimes(1);
+
+    expect(createSvgFilesMock).toBeCalledTimes(1);
+
+    expect(saveSummaryMock).toBeCalledTimes(0);
+
+    expect(mapAnswersToDtoMock).toBeCalledTimes(1);
+
+    expect(mockFillNullsForHiddenItems).toBeCalledTimes(1);
+
+    expect(getItemIdsMock).toBeCalledTimes(1);
+
+    expect(getFlowRecordMock).toBeCalledTimes(1);
+
+    expect(getUserIdentifierMock).toBeCalledTimes(1);
+
+    expect(mapUserActionsToDtoMock).toBeCalledTimes(1);
+
+    expect(mapAnswersToAlertsMock).toBeCalledTimes(1);
+
+    expect(getClientInformationMock).toBeCalledTimes(1);
+
+    expect(clearActivityStorageRecordMock).toBeCalledTimes(1);
+
+    expect(pushMock).toBeCalledTimes(1);
+
+    expect(pushMock).toBeCalledWith({
+      activityId: 'mock-activity-id-1',
+      alerts: [],
+      answers: [{ value: 3 }, { value: 8 }],
+      appletEncryption: 'applet-encryption-mock-1',
+      appletId: 'mock-applet-id-1',
+      client: {
+        appId: 'mock-app-name-1',
+        appVersion: 'mock-app-version-1',
+        height: 2358,
+        width: 1347,
+      },
+      createdAt: 1680956820000,
+      endTime: 1680956820000,
+      eventId: 'mock-event-id-1',
+      executionGroupKey: 'mock-flow-group-key-1',
+      flowId: 'mock-flow-id-1',
+      isFlowCompleted: false,
+      itemIds: ['mock-slider-id-1', 'mock-slider-id-2'],
+      logActivityName: 'mock-activity-name-1',
+      logCompletedAt: 'Sat, 08 Apr 2023 12:27:00 GMT',
+      scheduledTime: 1245800000,
+      startTime: 12389100000,
+      tzOffset: 3,
+      userActions: [],
+      userIdentifier: 'mock-user-id-1',
+      version: 'applet-version-mock-1',
+    });
+  });
+
+  it('Should push item to queue when hasSummary is true', async () => {
+    const answersMock: Answers = {
+      0: { answer: 3 },
+      1: { answer: 8 },
+    };
+
+    const itemsMock: PipelineItem[] = [
+      getSliderItem('mock-slider-item-1', 'mock-slider-id-1'),
+      getSliderItem('mock-slider-item-2', 'mock-slider-id-2'),
+    ];
+
+    const userActionsMock: UserAction[] = getUserActionsMock(answersMock);
+
+    const recordMock = getActivityRecordMockResult(
+      answersMock,
+      userActionsMock,
+      itemsMock,
+    )!;
+
+    jest.spyOn(storageHelpers, 'getActivityRecord').mockReturnValue({
+      ...recordMock,
+      hasSummary: true,
+    });
+
+    const pushMock = jest.fn();
+
+    const pushToQueueMock = { push: pushMock };
+
+    const dispatchMock = jest.fn();
+
+    const progress: StoreProgress = getFlowProgressMock();
+
+    const service = new ConstructCompletionsService(
+      saveSummaryMock,
+      {} as any,
+      progress,
+      pushToQueueMock,
+      dispatchMock,
+    );
+
+    //@ts-expect-error
+    service.queryDataUtils.getAppletDto = jest.fn().mockReturnValue({
+      encryption: 'applet-encryption-mock-1',
+      displayName: 'mock-applet-name-1',
+    });
+
+    await service.construct({
+      completionType: 'intermediate',
+      activityId: 'mock-activity-id-1',
+      activityName: 'mock-activity-name-1',
+      appletId: 'mock-applet-id-1',
+      eventId: 'mock-event-id-1',
+      flowId: 'mock-flow-id-1',
+      order: 0,
+    });
+
+    expect(saveSummaryMock).toBeCalledTimes(1);
+
+    expect(saveSummaryMock).toBeCalledWith({
+      activityId: 'mock-activity-id-1',
+      alerts: [
+        { activityItemId: 'mock-activity-item-1', message: 'mock-message-1' },
+      ],
+      order: 0,
+      scores: {
+        activityName: 'mock-activity-name-1',
+        scores: [{ flagged: true, name: 'mock-score-name-1', value: 125 }],
+      },
+    });
+
+    expect(pushMock).toBeCalledTimes(1);
+  });
+});
+
+describe('Test ConstructCompletionsService.constructForFinish', () => {
+  afterEach(() => {
+    jest.clearAllMocks();
+  });
+
+  it('Should push item to queue for activity flow', async () => {
+    const answersMock: Answers = {
+      0: { answer: 3 },
+      1: { answer: 8 },
+    };
+
+    const itemsMock: PipelineItem[] = [
+      getSliderItem('mock-slider-item-1', 'mock-slider-id-1'),
+      getSliderItem('mock-slider-item-2', 'mock-slider-id-2'),
+    ];
+
+    const userActionsMock: UserAction[] = getUserActionsMock(answersMock);
+
+    const getActivityRecordMock = jest
+      .spyOn(storageHelpers, 'getActivityRecord')
+      .mockReturnValue(
+        getActivityRecordMockResult(answersMock, userActionsMock, itemsMock),
+      );
+
+    const pushMock = jest.fn();
+
+    const pushToQueueMock = { push: pushMock };
+
+    const dispatchMock = jest.fn();
+
+    const progress: StoreProgress = getFlowProgressMock();
+
+    const service = new ConstructCompletionsService(
+      saveSummaryMock,
+      {} as any,
+      progress,
+      pushToQueueMock,
+      dispatchMock,
+    );
+
+    //@ts-expect-error
+    service.queryDataUtils.getAppletDto = jest.fn().mockReturnValue({
+      encryption: 'applet-encryption-mock-1',
+      displayName: 'mock-applet-name-1',
+    });
+
+    await service.construct({
+      completionType: 'finish',
+      activityId: 'mock-activity-id-1',
+      activityName: 'mock-activity-name-1',
+      appletId: 'mock-applet-id-1',
+      eventId: 'mock-event-id-1',
+      flowId: 'mock-flow-id-1',
+      order: 0,
+    });
+
+    expect(getActivityRecordMock).toBeCalledTimes(1);
+
+    expect(createSvgFilesMock).toBeCalledTimes(1);
+
+    expect(saveSummaryMock).toBeCalledTimes(0);
+
+    expect(mapAnswersToDtoMock).toBeCalledTimes(1);
+
+    expect(mockFillNullsForHiddenItems).toBeCalledTimes(1);
+
+    expect(getItemIdsMock).toBeCalledTimes(1);
+
+    expect(getFlowRecordMock).toBeCalledTimes(1);
+
+    expect(getUserIdentifierMock).toBeCalledTimes(1);
+
+    expect(mapUserActionsToDtoMock).toBeCalledTimes(1);
+
+    expect(mapAnswersToAlertsMock).toBeCalledTimes(1);
+
+    expect(getClientInformationMock).toBeCalledTimes(1);
+
+    expect(clearActivityStorageRecordMock).toBeCalledTimes(1);
+
+    expect(pushMock).toBeCalledTimes(1);
+
+    expect(pushMock).toBeCalledWith({
+      activityId: 'mock-activity-id-1',
+      alerts: [],
+      answers: [{ value: 3 }, { value: 8 }],
+      appletEncryption: 'applet-encryption-mock-1',
+      appletId: 'mock-applet-id-1',
+      client: {
+        appId: 'mock-app-name-1',
+        appVersion: 'mock-app-version-1',
+        height: 2358,
+        width: 1347,
+      },
+      createdAt: 1680956820000,
+      endTime: 1680956820000,
+      eventId: 'mock-event-id-1',
+      executionGroupKey: 'mock-flow-group-key-1',
+      flowId: 'mock-flow-id-1',
+      isFlowCompleted: true,
+      itemIds: ['mock-slider-id-1', 'mock-slider-id-2'],
+      logActivityName: 'mock-activity-name-1',
+      logCompletedAt: 'Sat, 08 Apr 2023 12:27:00 GMT',
+      scheduledTime: 1245800000,
+      startTime: 12389100000,
+      tzOffset: 3,
+      userActions: [],
+      userIdentifier: 'mock-user-id-1',
+      version: 'applet-version-mock-1',
+    });
+  });
+
+  it('Should push item to queue for a single activity', async () => {
+    const answersMock: Answers = {
+      0: { answer: 3 },
+      1: { answer: 8 },
+    };
+
+    const itemsMock: PipelineItem[] = [
+      getSliderItem('mock-slider-item-1', 'mock-slider-id-1'),
+      getSliderItem('mock-slider-item-2', 'mock-slider-id-2'),
+    ];
+
+    const userActionsMock: UserAction[] = getUserActionsMock(answersMock);
+
+    const getActivityRecordMock = jest
+      .spyOn(storageHelpers, 'getActivityRecord')
+      .mockReturnValue(
+        getActivityRecordMockResult(answersMock, userActionsMock, itemsMock),
+      );
+
+    const mockGetExecutionGroupKey = jest
+      .spyOn(operations, 'getExecutionGroupKey')
+      .mockReturnValue('mock-group-key-1');
+
+    const pushMock = jest.fn();
+
+    const pushToQueueMock = { push: pushMock };
+
+    const dispatchMock = jest.fn();
+
+    const progress: StoreProgress = getActivityProgressMock();
+
+    const service = new ConstructCompletionsService(
+      saveSummaryMock,
+      {} as any,
+      progress,
+      pushToQueueMock,
+      dispatchMock,
+    );
+
+    //@ts-expect-error
+    service.queryDataUtils.getAppletDto = jest.fn().mockReturnValue({
+      encryption: 'applet-encryption-mock-1',
+      displayName: 'mock-applet-name-1',
+    });
+
+    await service.construct({
+      completionType: 'finish',
+      activityId: 'mock-activity-id-1',
+      activityName: 'mock-activity-name-1',
+      appletId: 'mock-applet-id-1',
+      eventId: 'mock-event-id-1',
+      flowId: undefined,
+      order: 0,
+    });
+
+    expect(getActivityRecordMock).toBeCalledTimes(1);
+
+    expect(createSvgFilesMock).toBeCalledTimes(1);
+
+    expect(saveSummaryMock).toBeCalledTimes(0);
+
+    expect(mapAnswersToDtoMock).toBeCalledTimes(1);
+
+    expect(mockFillNullsForHiddenItems).toBeCalledTimes(1);
+
+    expect(getItemIdsMock).toBeCalledTimes(1);
+
+    expect(getFlowRecordMock).toBeCalledTimes(1);
+
+    expect(getUserIdentifierMock).toBeCalledTimes(1);
+
+    expect(mapUserActionsToDtoMock).toBeCalledTimes(1);
+
+    expect(mapAnswersToAlertsMock).toBeCalledTimes(1);
+
+    expect(getClientInformationMock).toBeCalledTimes(1);
+
+    expect(clearActivityStorageRecordMock).toBeCalledTimes(1);
+
+    expect(mockGetExecutionGroupKey).toBeCalledTimes(1);
+
+    expect(pushMock).toBeCalledTimes(1);
+
+    expect(pushMock).toBeCalledWith({
+      activityId: 'mock-activity-id-1',
+      alerts: [],
+      answers: [{ value: 3 }, { value: 8 }],
+      appletEncryption: 'applet-encryption-mock-1',
+      appletId: 'mock-applet-id-1',
+      client: {
+        appId: 'mock-app-name-1',
+        appVersion: 'mock-app-version-1',
+        height: 2358,
+        width: 1347,
+      },
+      createdAt: 1680956820000,
+      endTime: 1680956820000,
+      eventId: 'mock-event-id-1',
+      executionGroupKey: 'mock-group-key-1',
+      flowId: null,
+      isFlowCompleted: false,
+      itemIds: ['mock-slider-id-1', 'mock-slider-id-2'],
+      logActivityName: 'mock-activity-name-1',
+      logCompletedAt: 'Sat, 08 Apr 2023 12:27:00 GMT',
+      scheduledTime: 1245800000,
+      startTime: 12367800000,
+      tzOffset: 3,
+      userActions: [],
+      userIdentifier: 'mock-user-id-1',
+      version: 'applet-version-mock-1',
+    });
+  });
+});
+
+describe('Test ConstructCompletionsService: edge cases', () => {
+  it('"getAppletProperties" should throw error when no applet dto in the cache', () => {
+    const progress: StoreProgress = getFlowProgressMock();
+
+    const pushMock = jest.fn();
+
+    const pushToQueueMock = { push: pushMock };
+
+    const dispatchMock = jest.fn();
+
+    const service = new ConstructCompletionsService(
+      saveSummaryMock,
+      {} as any,
+      progress,
+      pushToQueueMock,
+      dispatchMock,
+    );
+
+    //@ts-expect-error
+    service.queryDataUtils.getAppletDto = jest.fn().mockReturnValue(null);
+
+    //@ts-expect-error
+    expect(() => service.getAppletProperties('mock-not-exist-id')).toThrow(
+      new Error(
+        "[ConstructCompletionsService] Applet doesn't exist in the cache, appletId=mock-not-exist-id ",
+      ),
+    );
+  });
+
+  const tests = [
+    {
+      activityStorageRecord: null,
+      appletEncryption: {},
+      expectedError:
+        '[ConstructCompletionsService] activityStorageRecord does not exist',
+    },
+    {
+      activityStorageRecord: undefined,
+      appletEncryption: {},
+      expectedError:
+        '[ConstructCompletionsService] activityStorageRecord does not exist',
+    },
+    {
+      activityStorageRecord: {},
+      appletEncryption: null,
+      expectedError:
+        '[ConstructCompletionsService] Encryption params is undefined',
+    },
+    {
+      activityStorageRecord: {},
+      appletEncryption: undefined,
+      expectedError:
+        '[ConstructCompletionsService] Encryption params is undefined',
+    },
+  ];
+
+  tests.forEach(
+    ({ activityStorageRecord, appletEncryption, expectedError }) => {
+      const activityStorageRecordAsText = activityStorageRecord
+        ? 'fullfilled'
+        : activityStorageRecord;
+      const appletEncryptionAsText = appletEncryption
+        ? 'fullfilled'
+        : appletEncryption;
+
+      it(`"validate" should throw error when activityStorageRecord is ${activityStorageRecordAsText} and appletEncryption is ${appletEncryptionAsText}`, () => {
+        const progress: StoreProgress = getFlowProgressMock();
+
+        const pushMock = jest.fn();
+
+        const pushToQueueMock = { push: pushMock };
+
+        const dispatchMock = jest.fn();
+
+        const service = new ConstructCompletionsService(
+          saveSummaryMock,
+          {} as any,
+          progress,
+          pushToQueueMock,
+          dispatchMock,
+        );
+
+        expect(() =>
+          //@ts-expect-error
+          service.validate(
+            activityStorageRecord as any,
+            appletEncryption as any,
+          ),
+        ).toThrow(new Error(expectedError));
+      });
+    },
+  );
+});
